requirements.txt
google-generativeai
networkx
python-dotenv
fastapi
pillow
uvicorn